+++ aten/src/ATen/templates/TensorMethods.h
@@ -4,7 +4,7 @@
 #include <c10/core/MemoryFormat.h>
 #include <c10/macros/Macros.h>
 #include <c10/core/TensorOptions.h>
-#include <ATen/core/DeprecatedTypeProperties.h>
+#include <ATen/DeprecatedTypeProperties.h>
 
 namespace at {
